# Conflicts:
#	core/src/Framework/FWTrajectoryPlanner.cpp

+++ core/src/Framework/FWStaticTorqueOptimizer.cpp
@@ -134,13 +134,13 @@ void GrahamConvexHull::Recalc(std::vector<Vec3f> vertices) {
 		vertices[0] = vertices[xmax];
 		vertices[xmax] = tmp;
 		DSTR << "before sort" << std::endl;
-		for (int i = 0; i < vertices.size(); i++) {
+		for (int i = 0; i < (int)vertices.size(); i++) {
 			DSTR << vertices[i] << std::endl;
 		}
 		Sort::simplesort(vertices, normal);
 		//Sort::quicksort(vertices, 1, n - 1, normal);
 		DSTR << "after sort" << std::endl;
-		for (int i = 0; i < vertices.size(); i++) {
+		for (int i = 0; i < (int)vertices.size(); i++) {
 			DSTR << vertices[i] << std::endl;
 		}
 		vertices.push_back(vertices[0]);
@@ -725,7 +725,7 @@ double FWStaticTorqueOptimizer::CalcTorqueInChildren(PHIKActuatorIf* root, Vec3d
 		force += forceInChildren;
 	}
 	//Ú’n„‘Ì‚ÌR—Í‚Ì‡¬
-	for (int i = 0; i < groundConst.size(); i++) {
+	for (int i = 0; i < (int)groundConst.size(); i++) {
 		if (rootSolid == groundConst[i]->cSolid) {
 			DSTR << "match : " << rootSolid->GetName() << "&groundConst[" << i << "]" << std::endl;
 			double t = groundConst[i]->contactForce.norm() / (force.norm() + groundConst[i]->contactForce.norm());
